feat: sweet alert dan multi user

+++ page/peminjaman_detail/index.php
@@ -16,7 +16,7 @@
       <div class="card-body ">
         <a href="index.php?page=peminjaman_detail&act=create" class="btn btn-info mb-3">TAMBAH DATA</a>
 
-        <table class="table table-bordered table-hover table-sm">
+        <table class="table table-striped table-bordered table-hover text-center" id="dataTables-example">
           <thead class="thead-light">
             <tr class="text-center">
               <th scope="col" class="font-weight-bold">Id Peminjaman</th>
@@ -36,19 +36,30 @@
             $peminjaman_details = $peminjaman_detail->getAll();
 
             foreach ($peminjaman_details as $row) {
-            ?>
+              ?>
               <tr>
-                <td class="text-center"><?php echo htmlspecialchars($row['id_peminjaman']); ?></td>
-                <td><?php echo htmlspecialchars($row['judul']); ?></td>
-                <td><?php echo htmlspecialchars($row['denda']); ?></td>
+                <td class="text-center">
+                  <?php echo htmlspecialchars($row['id_peminjaman']); ?>
+                </td>
+                <td>
+                  <?php echo htmlspecialchars($row['judul']); ?>
+                </td>
+                <td>
+                  <?php echo htmlspecialchars($row['denda']); ?>
+                </td>
                 <td class="text-center">
                   <a href="index.php?page=peminjaman_detail&act=update&id_peminjaman_detail=<?php echo $row['id_peminjaman_detail'] ?>"
                     class="btn btn-info btn-sm">Edit</a>
-                  <a href="index.php?page=peminjaman_detail&act=delete&id_peminjaman_detail=<?php echo $row['id_peminjaman_detail'] ?>"
-                    class="btn btn-danger btn-sm">Hapus</a>
+
+                  <?php if (isset($_SESSION['role']) && $_SESSION['role'] === 'super_admin') { ?>
+                    <a href="index.php?page=peminjaman_detail&act=delete&id_peminjaman_detail=<?php echo $row['id_peminjaman_detail'] ?>"
+                      class="btn btn-danger btn-sm">Hapus</a>
+                    <?php
+                  }
+                  ?>
                 </td>
               </tr>
-            <?php
+              <?php
             }
             dataBase::disconnect();
             ?>
