+++ page/penulis/index.php
@@ -16,7 +16,7 @@
       <div class="card-body">
         <a href="index.php?page=penulis&act=create" class="btn btn-info mb-3">TAMBAH DATA</a>
 
-        <table class="table table-bordered table-hover table-sm">
+        <table class="table table-striped table-bordered table-hover text-center" id="dataTables-example">
           <thead class="thead-light">
             <tr class="text-center">
               <th scope="col" class="font-weight-bold">No</th>
@@ -38,20 +38,33 @@
             $no = 1;
 
             foreach ($datapenulis as $row) {
-            ?>
+              ?>
               <tr>
-                <td class="text-center"><?php echo $no++ ?></td>
-                <td><?php echo htmlspecialchars($row['nama_penulis']) ?></td>
-                <td><?php echo htmlspecialchars($row['asal_negara']) ?></td>
-                <td><?php echo htmlspecialchars($row['jumlah_karya']) ?></td>
+                <td class="text-center">
+                  <?php echo $no++ ?>
+                </td>
+                <td>
+                  <?php echo htmlspecialchars($row['nama_penulis']) ?>
+                </td>
+                <td>
+                  <?php echo htmlspecialchars($row['asal_negara']) ?>
+                </td>
+                <td>
+                  <?php echo htmlspecialchars($row['jumlah_karya']) ?>
+                </td>
                 <td class="text-center">
                   <a href="index.php?page=penulis&act=update&id_penulis=<?php echo $row['id_penulis'] ?>"
                     class="btn btn-primary btn-sm">Edit</a>
-                  <a href="index.php?page=penulis&act=delete&id_penulis=<?php echo $row['id_penulis'] ?>"
-                    class="btn btn-danger btn-sm">Hapus</a>
+
+                  <?php if (isset($_SESSION['role']) && $_SESSION['role'] === 'super_admin') { ?>
+                    <a href="index.php?page=penulis&act=delete&id_penulis=<?php echo $row['id_penulis'] ?>"
+                      class="btn btn-danger btn-sm">Hapus</a>
+                    <?php
+                  }
+                  ?>
                 </td>
               </tr>
-            <?php
+              <?php
             }
             dataBase::disconnect();
             ?>
